Setup HTML, add CSS file
In this commit, I've added HTML/CSS files to the repo, and set up the
basic content within the HTML to set up a baseline. Some of the ways I
am choosing to set up the content (in response to the design file in the
TOP prompt) will set me up to practice using a CSS reset.