Add/implement image assets
In this commit, I've added two image assets. One, an image of a light
show, has been implemented in the CSS as a background image for the
image-block class. The second, a png of The Odin Project's logo, is
contained directly in the HTML as part of the hero class.