Add styles/organization
In this commit, I've added additional divs, classes, and ids to the html
for better organization of the layout. I've also added some content to
the HTML to add a pattern validation against a regular expression to the
phone number input field. From there, I organized the layout and styles
using CSS, as well as fonts linked using the Google Fonts API. At the
top of the stylesheet, I've included the meyerweb reset. In future
commits, I will be adding some resetting CSS to neutralize the styling
of the form controls as well.