Add reset for form styling, style input fonts
In this commit, I've implemented the resets for form styling from
Normalize.css, and used comments in the CSS to label the different
sections for the meyerweb resets, Normalize.css resets, and the actual
styles for the page. With help from the Normalize.css resets, I've also
added font-family styles for the input fields.